cambios

+++ sigo/app/controllers/payrolls/payslips_controller.rb
@@ -65,8 +65,9 @@ class Payrolls::PayslipsController < ApplicationController
         render :new, layout: false 
       end
     end
-
-    ActiveRecord::Base.connection.execute("DELETE FROM extra_information_for_payslips WHERE week = '"+params[:extra_information_for_payslip][''+regsEx.first.to_s+'']['week'].to_s+"'" )
+    if regsEx.length < 1
+      ActiveRecord::Base.connection.execute("DELETE FROM extra_information_for_payslips WHERE week = '"+params[:extra_information_for_payslip][''+regsEx.first.to_s+'']['week'].to_s+"'" )
+    end
     regsEx.each do |reg|
       extra = ExtraInformationForPayslip.new
       extra.worker_id = params[:extra_information_for_payslip][''+reg.to_s+'']['worker_id']
